+++ src/main/java/sayTheSpire/ui/elements/CardElement.java
@@ -20,7 +20,7 @@ public class CardElement extends GameObjectElement {
 
     public enum CardLocation {
         HAND, GRID_SELECT, HAND_SELECT, MASTER_DECK_VIEW, EXHAUST_PILE_VIEW, DISCARD_PILE_VIEW, DRAW_PILE_VIEW, SHOP,
-        OTHER
+        COMPENDIUM, OTHER
     }
 
     protected AbstractCard card;